+++ lib/generators/sorcery/templates/initializer.rb
@@ -226,7 +226,7 @@ Rails.application.config.sorcery.configure do |config|
   # config.line.bot_prompt = "normal"
   # config.line.user_info_mapping = {name: 'displayName'}
 
-  
+
   # For information about Discord API
   # https://discordapp.com/developers/docs/topics/oauth2
   # config.discord.key = "xxxxxx"
@@ -362,8 +362,8 @@ Rails.application.config.sorcery.configure do |config|
     # user.activation_mailer_disabled =
 
     # Method to send email related
-    # options: `:deliver_later`, `:deliver_now`, `:deliver`
-    # Default: :deliver (Rails version < 4.2) or :deliver_now (Rails version 4.2+)
+    # options: `:deliver_later`, `:deliver_now`
+    # Default: :deliver_now
     #
     # user.email_delivery_method =