+++ spec/rails_app/db/migrate/activity_logging/20101224223624_add_activity_logging_to_users.rb
@@ -1,4 +1,4 @@
-class AddActivityLoggingToUsers < ActiveRecord::CompatibleLegacyMigration.migration_class
+class AddActivityLoggingToUsers < ActiveRecord::Migration::Current
   def self.up
     add_column :users, :last_login_at,     :datetime, default: nil
     add_column :users, :last_logout_at,    :datetime, default: nil