+++ spec/rails_app/db/migrate/invalidate_active_sessions/20180221093235_add_invalidate_active_sessions_before_to_users.rb
@@ -1,4 +1,4 @@
-class AddInvalidateSessionsBeforeToUsers < ActiveRecord::CompatibleLegacyMigration.migration_class
+class AddInvalidateSessionsBeforeToUsers < ActiveRecord::Migration::Current
   def self.up
     add_column :users, :invalidate_sessions_before, :datetime, default: nil
   end