+++ spec/rails_app/db/migrate/remember_me/20101224223623_add_remember_me_token_to_users.rb
@@ -1,4 +1,4 @@
-class AddRememberMeTokenToUsers < ActiveRecord::CompatibleLegacyMigration.migration_class
+class AddRememberMeTokenToUsers < ActiveRecord::Migration::Current
   def self.up
     add_column :users, :remember_me_token, :string, default: nil
     add_column :users, :remember_me_token_expires_at, :datetime, default: nil